Project 5
Improvements: A log file is appended at each function call and a csv file is created at the end of the execution to save users and account data. At the beginning of the next execution, the csv file is read at the information is extracted to recreate the previous user and account list. For more information, read READ.ME.